refactor the countdown interval to use Date.now
I didn't realize how inaccurate the countdown timer
was until I used it at an event. I put 10 minutes on it
and it was slow by 3 minutes.
I refactored the code to use Date.now, which takes
time from the system's clock, and I set polling to
happen every 100ms